+++ test/helpers/getCleanTestingRoot.js
@@ -1,19 +1,7 @@
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 
 export function getCleanTestingRoot() {
-  let $root = document.getElementById('root');
+  const node = document.getElementById('root');
 
-  // if a root exists, unmount anything inside and remove it
-  if ($root) {
-    ReactDOM.unmountComponentAtNode($root);
-    $root.parentNode.removeChild($root);
-  }
-
-  // create a brand new root element
-  $root = document.createElement('div');
-  $root.id = 'root';
-
-  document.body.appendChild($root);
-
-  return $root;
+  return createRoot(node);
 }